pcre32: compile: Fix signed/unsigned mismatch in check_posix_name
git-svn-id: svn://vcs.exim.org/pcre/code/trunk@1074 2f5784b3-3f2a-0410-8824-cb99058d5e15

+++ pcre_compile.c
@@ -2755,7 +2755,7 @@ register int yield = 0;
 while (posix_name_lengths[yield] != 0)
   {
   if (len == posix_name_lengths[yield] &&
-    STRNCMP_UC_C8(ptr, pn, len) == 0) return yield;
+    STRNCMP_UC_C8(ptr, pn, (unsigned int)len) == 0) return yield;
   pn += posix_name_lengths[yield] + 1;
   yield++;
   }
